requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
loguru>=0.7.2
python-multipart>=0.0.9
PyMuPDF>=1.24.3
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23
sentence-transformers>=2.7.0
transformers>=4.41.0
torch>=2.2.0
faiss-cpu>=1.7.4
ollama>=0.1.9
httpx>=0.27.0
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.9.0
seaborn>=0.13.2
Pillow>=10.3.0
tqdm>=4.66.4
graphviz